Split oversized werewolf player-night-state and owner-state services
Extract cohesive concerns from the two over-cap service files into
sibling modules, keeping the entry functions as thin orchestrators and
public exports stable.

player-night-state.ts (626 -> 211): extract per-role solo dispatch to
player-night-state-roles.ts, seating/positional roles (The Thing,
Insomniac, Count) to player-night-state-positional.ts, and interim-attack
roles (Witch, Altruist) to player-night-state-protection.ts.

owner-state.ts (420 -> 99): extract the daytime night-outcome summary to
owner-daytime-summary.ts, daytime trial/nomination state to
owner-daytime-state.ts, and the small turn-state extractors to
owner-state-helpers.ts. Re-export moved symbols so importers stay stable.

No behavior change; the full suite stays green.
