+++ src/lib/game/modes/werewolf/services/owner-state.ts
@@ -1,342 +1,16 @@
 import type { Game } from "@/lib/types";
 import { GameMode } from "@/lib/types";
-import type { DaytimeNightStatusEntry } from "@/server/types";
 
-import { getWerewolfModeConfig } from "../lobby-config";
 import type { WerewolfPlayerGameState } from "../player-state";
-import {
-  getWerewolfRole,
-  isWerewolfRole,
-  WEREWOLF_ROLES,
-  WerewolfRole,
-} from "../roles";
-import type {
-  AltruistInterceptedNightResolutionEvent,
-  AnyNightAction,
-} from "../types";
-import { TrialVerdict, WerewolfPhase } from "../types";
-import { OLD_MAN_TIMER_KEY, SMITE_PHASE_KEY } from "../utils";
-import { getHypnotizedPlayerId, getSilencedPlayerIds } from "../utils";
+import { WerewolfPhase } from "../types";
 import { currentTurnState } from "../utils/game-state";
-
-/** Returns the set of player IDs whose night outcomes are publicly visible. */
-function resolveRevealedPlayerIds(game: Game): Set<string> {
-  const ts = currentTurnState(game);
-  if (ts?.phase.type !== WerewolfPhase.Daytime) return new Set();
-  return new Set(ts.phase.revealedPlayerIds ?? []);
-}
-
-/** Extracts nightActions from the current turnState, if present. */
-function extractNightActions(
-  game: Game,
-): Record<string, AnyNightAction> | undefined {
-  const ts = currentTurnState(game);
-  if (!ts) return undefined;
-  const { nightActions } = ts.phase;
-  if (ts.phase.type === WerewolfPhase.Nighttime) return nightActions;
-  return Object.keys(nightActions).length > 0 ? nightActions : undefined;
-}
-
-/** Extracts deadPlayerIds from the Werewolf turn state. */
-function extractDeadPlayerIds(game: Game): string[] {
-  const ts = currentTurnState(game);
-  return ts?.deadPlayerIds ?? [];
-}
-
-/** Extracts the Hunter revenge player ID (narrator-only). */
-function extractHunterRevengePlayerId(game: Game): string | undefined {
-  const ts = currentTurnState(game);
-  return ts?.roleState?.hunter?.revengePlayerId;
-}
-
-/**
- * Extracts sanitized night outcomes for the daytime summary.
- * Shared between owner and player views.
- */
-export function extractDaytimeNightSummary(
-  game: Game,
-  callerId: string,
-): Partial<WerewolfPlayerGameState> {
-  const ts = currentTurnState(game);
-  if (ts?.phase.type !== WerewolfPhase.Daytime) return {};
-  const phase = ts.phase;
-  const revealedPlayerIds = resolveRevealedPlayerIds(game);
-  const callerIsOwner = callerId === game.ownerPlayerId;
-
-  const altruistIntercept = (phase.nightResolution ?? []).find(
-    (e): e is AltruistInterceptedNightResolutionEvent =>
-      e.type === "altruist-intercepted",
-  );
-
-  const canSeeOutcomeFor = (playerId: string) =>
-    callerIsOwner || revealedPlayerIds.has(playerId) || playerId === callerId;
-
-  const nightStatus: DaytimeNightStatusEntry[] = (
-    phase.nightResolution ?? []
-  ).flatMap((e): DaytimeNightStatusEntry[] => {
-    if (e.type === "killed" && e.died) {
-      if (!canSeeOutcomeFor(e.targetPlayerId)) return [];
-      // Altruist death: emit altruist-sacrifice with savedPlayerId.
-      if (e.targetPlayerId === altruistIntercept?.altruistPlayerId) {
-        return [
-          {
-            targetPlayerId: e.targetPlayerId,
-            effect: "altruist-sacrifice",
-            savedPlayerId: altruistIntercept.savedPlayerId,
-          },
-        ];
-      }
-      if (e.attackedBy.includes(SMITE_PHASE_KEY)) {
-        return [{ targetPlayerId: e.targetPlayerId, effect: "smited" }];
-      }
-      const effect =
-        e.attackedBy.length === 1 && e.attackedBy[0] === OLD_MAN_TIMER_KEY
-          ? "peaceful"
-          : "killed";
-      return [{ targetPlayerId: e.targetPlayerId, effect }];
-    }
-    // Attacked but saved by protection — visible to all when setting is on.
-    if (
-      e.type === "killed" &&
-      !e.died &&
-      e.protectedBy.length > 0 &&
-      getWerewolfModeConfig(game).revealProtections
-    ) {
-      return [{ targetPlayerId: e.targetPlayerId, effect: "protected" }];
-    }
-    if (e.type === "tough-guy-absorbed" && e.targetPlayerId === callerId)
-      return [{ targetPlayerId: e.targetPlayerId, effect: "survived" }];
-    if (e.type === "silenced" && canSeeOutcomeFor(e.targetPlayerId))
-      return [{ targetPlayerId: e.targetPlayerId, effect: "silenced" }];
-    if (e.type === "hypnotized" && canSeeOutcomeFor(e.targetPlayerId))
-      return [{ targetPlayerId: e.targetPlayerId, effect: "hypnotized" }];
-    // Veteran counter-kill: emit when visible and the counter-killed player died.
-    if (
-      e.type === "veteran-counterkilled" &&
-      e.died &&
-      canSeeOutcomeFor(e.counterkilledPlayerId)
-    ) {
-      return [
-        {
-          targetPlayerId: e.counterkilledPlayerId,
-          effect: "veteran-counterkill",
-          veteranPlayerId: e.veteranPlayerId,
-          veteranCounterkillSource: e.source,
-        },
-      ];
-    }
-    return [];
-  });
-  if (phase.knightedPlayerId !== undefined) {
-    nightStatus.push({
-      targetPlayerId: phase.knightedPlayerId,
-      effect: "knighted",
-    });
-  }
-
-  // Exposer reveal: emit an "exposed" entry visible to all players. The reveal
-  // lives on the daytime phase so it only surfaces on the day after the
-  // exposure, not on every subsequent day.
-  if (phase.exposerReveal) {
-    const exposerReveal = phase.exposerReveal;
-    const revealedPlayer = game.players.find(
-      (p) => p.id === exposerReveal.playerId,
-    );
-    const revealedRoleDef = getWerewolfRole(exposerReveal.roleId);
-    if (revealedPlayer && revealedRoleDef) {
-      nightStatus.push({
-        targetPlayerId: exposerReveal.playerId,
-        effect: "exposed",
-        roleName: revealedRoleDef.name,
-      });
-    }
-  }
-
-  return {
-    ...(nightStatus.length > 0 ? { nightStatus } : {}),
-  };
-}
-
-/**
- * Extracts daytime trial and nomination state visible to a specific player.
- */
-export function extractDaytimePlayerState(
-  game: Game,
-  callerId: string,
-): Partial<WerewolfPlayerGameState> {
-  const ts = currentTurnState(game);
-  if (ts?.phase.type !== WerewolfPhase.Daytime) return {};
-  const phase = ts.phase;
-
-  const result: Partial<WerewolfPlayerGameState> = {};
-
-  // Nominations.
-  if (getWerewolfModeConfig(game).nominationsEnabled) {
-    const nominations = phase.nominations ?? [];
-    const nominatorsByDefendant = nominations.reduce<Record<string, string[]>>(
-      (acc, n) => {
-        (acc[n.defendantId] ??= []).push(n.nominatorId);
-        return acc;
-      },
-      {},
-    );
-    result.nominations = Object.entries(nominatorsByDefendant).map(
-      ([defendantId, nominatorIds]) => ({ defendantId, nominatorIds }),
-    );
-    const myNomination = nominations.find((n) => n.nominatorId === callerId);
-    if (myNomination) {
-      result.myNominatedDefendantId = myNomination.defendantId;
-    }
-  }
-
-  // Narrator-only: pending daytime smites.
-  if (callerId === game.ownerPlayerId && phase.pendingSmitePlayerIds?.length) {
-    result.pendingSmitePlayerIds = phase.pendingSmitePlayerIds;
-  }
-
-  // Executioner target.
-  const callerExecutionerAssignment = game.roleAssignments.find(
-    (a) =>
-      a.playerId === callerId &&
-      a.roleDefinitionId === (WerewolfRole.Executioner as string),
-  );
-  if (callerExecutionerAssignment && ts.roleState?.executioner?.targetId) {
-    result.executionerTargetId = ts.roleState.executioner.targetId;
-  }
-
-  // The Thing tap notification.
-  if (ts.roleState?.theThing?.tapped === callerId) result.thingTappedMe = true;
-
-  // Silenced / hypnotized.
-  const silencedIds = getSilencedPlayerIds(ts);
-  if (silencedIds.includes(callerId)) result.isSilenced = true;
-  if (getHypnotizedPlayerId(ts) === callerId) result.isHypnotized = true;
-
-  // Active trial.
-  if (phase.activeTrial) {
-    const { activeTrial } = phase;
-    const alivePlayerCount = game.players.filter(
-      (p) =>
-        p.id !== game.ownerPlayerId &&
-        p.id !== activeTrial.defendantId &&
-        !ts.deadPlayerIds.includes(p.id) &&
-        !silencedIds.includes(p.id),
-    ).length;
-    const myVote = activeTrial.votes.find((v) => v.playerId === callerId)?.vote;
-    const playerById = new Map(game.players.map((p) => [p.id, p]));
-
-    const callerRoleId = game.roleAssignments.find(
-      (a) => a.playerId === callerId,
-    )?.roleDefinitionId;
-    const mustVoteGuilty =
-      callerRoleId !== undefined &&
-      isWerewolfRole(callerRoleId) &&
-      WEREWOLF_ROLES[callerRoleId].alwaysVotesGuilty === true;
-    const mustVoteInnocent =
-      callerRoleId !== undefined &&
-      isWerewolfRole(callerRoleId) &&
-      WEREWOLF_ROLES[callerRoleId].alwaysVotesInnocent === true;
-
-    result.activeTrial = {
-      defendantId: activeTrial.defendantId,
-      startedAt: activeTrial.startedAt,
-      phase: activeTrial.phase,
-      ...(activeTrial.voteStartedAt !== undefined
-        ? { voteStartedAt: activeTrial.voteStartedAt }
-        : {}),
-      ...(activeTrial.pausedAt !== undefined
-        ? { pausedAt: activeTrial.pausedAt }
-        : {}),
-      ...(activeTrial.pauseOffset !== undefined
-        ? { pauseOffset: activeTrial.pauseOffset }
-        : {}),
-      ...(myVote !== undefined ? { myVote } : {}),
-      voteCount: activeTrial.votes.length,
-      playerCount: alivePlayerCount,
-      ...(activeTrial.verdict ? { verdict: activeTrial.verdict } : {}),
-      ...(mustVoteGuilty ? { mustVoteGuilty: true } : {}),
-      ...(mustVoteInnocent ? { mustVoteInnocent: true } : {}),
-    };
-
-    if (activeTrial.verdict) {
-      result.activeTrial.voteResults = activeTrial.votes.map((v) => ({
-        playerName: playerById.get(v.playerId)?.name ?? v.playerId,
-        vote: v.vote,
-      }));
-
-      // Expose whether the defendant was actually eliminated so both player
-      // and narrator panels can derive "spared" from the same source of truth.
-      // Only meaningful for Eliminated verdicts; leave undefined for Innocent.
-      if (activeTrial.verdict === TrialVerdict.Eliminated) {
-        result.activeTrial.defendantEliminated = ts.deadPlayerIds.includes(
-          activeTrial.defendantId,
-        );
-      }
-
-      // Suppress eliminatedRole if the defendant survived (e.g. Martyr intercept).
-      // Only reveal when the defendant is actually dead.
-      if (
-        activeTrial.verdict === TrialVerdict.Eliminated &&
-        ts.deadPlayerIds.includes(activeTrial.defendantId)
-      ) {
-        const assignment = game.roleAssignments.find(
-          (a) => a.playerId === activeTrial.defendantId,
-        );
-        const roleDef = assignment
-          ? getWerewolfRole(assignment.roleDefinitionId)
-          : undefined;
-        if (roleDef) {
-          result.activeTrial.eliminatedRole = {
-            id: roleDef.id,
-            name: roleDef.name,
-            team: roleDef.team,
-          };
-        }
-      }
-    }
-  }
-
-  // Concluded trials count for this day.
-  if (phase.concludedTrialsCount) {
-    result.concludedTrialsCount = phase.concludedTrialsCount;
-  }
-
-  // Martyr window: expose pendingGuiltId so the UI can display the window.
-  if (phase.pendingGuiltId) {
-    result.pendingGuiltId = phase.pendingGuiltId;
-  }
-
-  // Martyr ability used flag: only expose to the Martyr player so they know
-  // whether they can still intervene.
-  const callerMartyrAssignment = game.roleAssignments.find(
-    (a) =>
-      a.playerId === callerId &&
-      a.roleDefinitionId === (WerewolfRole.Martyr as string),
-  );
-  if (callerMartyrAssignment && ts.roleState?.martyr?.abilityUsed) {
-    result.martyrUsed = true;
-  }
-
-  // Ghost clues — visible to all players during daytime.
-  const ghostClues =
-    ts.roleState?.ghost?.clues.filter((c) => c.turn <= ts.turn) ?? [];
-  if (ghostClues.length > 0) {
-    result.ghostClues = ghostClues;
-  }
-
-  // Ghost clue submission status — only relevant for the Ghost player.
-  const callerAssignment = game.roleAssignments.find(
-    (a) => a.playerId === callerId,
-  );
-  if (callerAssignment?.roleDefinitionId === (WerewolfRole.Ghost as string)) {
-    const submittedThisTurn = (ts.roleState?.ghost?.clues ?? []).some(
-      (c) => c.turn === ts.turn,
-    );
-    result.ghostClueSubmittedThisTurn = submittedThisTurn;
-  }
-
-  return result;
-}
+import { extractDaytimeNightSummary } from "./owner-daytime-summary";
+import {
+  extractDeadPlayerIds,
+  extractHunterRevengePlayerId,
+  extractNightActions,
+  resolveRevealedPlayerIds,
+} from "./owner-state-helpers";
 
 /**
  * Extracts complete owner/narrator state from the Werewolf game.
@@ -382,8 +56,6 @@ export function extractOwnerState(
   };
 }
 
-export { extractDeadPlayerIds, extractNightActions };
-
 /**
  * Returns dead players visible to a specific caller.
  * During manual night-outcome reveal, newly killed players remain hidden from
@@ -418,3 +90,10 @@ export function extractVisibleDeadPlayerIds(
   }
   return visible;
 }
+
+export { extractDaytimePlayerState } from "./owner-daytime-state";
+export { extractDaytimeNightSummary } from "./owner-daytime-summary";
+export {
+  extractDeadPlayerIds,
+  extractNightActions,
+} from "./owner-state-helpers";
