+++ src/lib/game/modes/werewolf/services/player-night-state.ts
@@ -1,6 +1,4 @@
 import type { Game } from "@/lib/types";
-import { Team } from "@/lib/types";
-import type { NighttimeNightStatusEntry } from "@/server/types";
 
 import type { WerewolfPlayerGameState } from "../player-state";
 import type { WerewolfRoleDefinition } from "../roles";
@@ -11,46 +9,9 @@ import {
   baseGroupPhaseKey,
   currentTurnState,
   getGroupPhasePlayerIds,
-  getInterimAttackedPlayerIds,
 } from "../utils";
 import { appendInvestigationResult } from "./player-night-state-helpers";
-
-/**
- * Returns the player IDs immediately adjacent to `playerId` in the given
- * seating order. Returns up to two IDs (left and right neighbours), wrapping
- * around the ends of the list. If the player is not found, returns [].
- */
-function getAdjacentPlayerIds(
-  playerOrder: string[],
-  playerId: string,
-): string[] {
-  const idx = playerOrder.indexOf(playerId);
-  if (idx === -1 || playerOrder.length < 2) return [];
-  const left = playerOrder[(idx - 1 + playerOrder.length) % playerOrder.length];
-  const right = playerOrder[(idx + 1) % playerOrder.length];
-  if (!left || !right) return [];
-  if (left === right) return [left];
-  return [left, right];
-}
-
-/**
- * Returns the ID of the left neighbor (seat before) and right neighbor
- * (seat after) for `playerId` in `playerOrder`. Returns undefined for either
- * if the player list is too short.
- */
-function getNeighborIds(
-  playerOrder: string[],
-  playerId: string,
-): { left: string | undefined; right: string | undefined } {
-  const idx = playerOrder.indexOf(playerId);
-  if (idx === -1 || playerOrder.length < 2) {
-    return { left: undefined, right: undefined };
-  }
-  return {
-    left: playerOrder[(idx - 1 + playerOrder.length) % playerOrder.length],
-    right: playerOrder[(idx + 1) % playerOrder.length],
-  };
-}
+import { extractRoleSpecificState } from "./player-night-state-roles";
 
 function hasPriestActiveWard(ts: WerewolfTurnState | undefined): boolean {
   if (!ts?.roleState?.priest?.wards) return false;
@@ -197,382 +158,6 @@ function extractGroupPhaseState(
     ts?.roleState?.evilEmpath?.revealedResult !== undefined
       ? { evilEmpathRevealedResult: ts.roleState.evilEmpath.revealedResult }
       : {}),
-  };
-}
-
-function extractRoleSpecificState(
-  game: Game,
-  callerId: string,
-  myRole: WerewolfRoleDefinition,
-  nightActions: Record<string, AnyNightAction>,
-  deadPlayerIds: string[],
-  ts: WerewolfTurnState | undefined,
-): Partial<WerewolfPlayerGameState> | undefined {
-  if (myRole.id === WerewolfRole.Exposer) {
-    const exposerAction = nightActions[myRole.id];
-    const soloAction =
-      exposerAction && !isTeamNightAction(exposerAction)
-        ? exposerAction
-        : undefined;
-    return {
-      myNightTarget: soloAction?.skipped ? null : soloAction?.targetPlayerId,
-      myNightTargetConfirmed: soloAction?.confirmed ?? false,
-      exposerAbilityUsed: ts?.roleState?.exposer?.abilityUsed ?? false,
-    };
-  }
-
-  if (myRole.id === WerewolfRole.Mortician) {
-    const morticianAction = nightActions[myRole.id];
-    const soloAction =
-      morticianAction && !isTeamNightAction(morticianAction)
-        ? morticianAction
-        : undefined;
-    return {
-      myNightTarget: soloAction?.skipped ? null : soloAction?.targetPlayerId,
-      myNightTargetConfirmed: soloAction?.confirmed ?? false,
-      morticianAbilityEnded: ts?.roleState?.mortician?.abilityEnded ?? false,
-    };
-  }
-
-  if (myRole.id === WerewolfRole.Witch) {
-    return extractWitchState(game, nightActions, myRole, deadPlayerIds, ts);
-  }
-
-  if (myRole.id === WerewolfRole.Altruist) {
-    return extractAltruistState(
-      game,
-      callerId,
-      nightActions,
-      myRole,
-      deadPlayerIds,
-      ts,
-    );
-  }
-
-  if (myRole.id === WerewolfRole.Mirrorcaster) {
-    const mcAction = nightActions[myRole.id];
-    const soloAction =
-      mcAction && !isTeamNightAction(mcAction) ? mcAction : undefined;
-    return {
-      myNightTarget: soloAction?.skipped ? null : soloAction?.targetPlayerId,
-      myNightTargetConfirmed: soloAction?.confirmed ?? false,
-      mirrorcasterCharged: ts?.roleState?.mirrorcaster?.charged ?? false,
-    };
-  }
-
-  if (myRole.id === WerewolfRole.Mercenary) {
-    const mercAction = nightActions[myRole.id];
-    const soloAction =
-      mercAction && !isTeamNightAction(mercAction) ? mercAction : undefined;
-    return {
-      myNightTarget: soloAction?.skipped ? null : soloAction?.targetPlayerId,
-      myNightTargetConfirmed: soloAction?.confirmed ?? false,
-      mercenaryCharged: ts?.roleState?.mercenary?.charged ?? false,
-    };
-  }
-
-  if (myRole.id === WerewolfRole.Veteran) {
-    const vetAction = nightActions[myRole.id];
-    const soloAction =
-      vetAction && !isTeamNightAction(vetAction) ? vetAction : undefined;
-    return {
-      myNightTarget: soloAction?.skipped ? null : undefined,
-      myNightTargetConfirmed: soloAction?.confirmed ?? false,
-      veteranAlertsUsed: ts?.roleState?.veteran?.alertsUsed ?? 0,
-      myNightAlerted: soloAction?.alerted === true,
-    };
-  }
-
-  if (myRole.id === WerewolfRole.Executioner) {
-    return {
-      myNightTarget: undefined,
-      myNightTargetConfirmed: false,
-      ...(ts?.roleState?.executioner?.targetId
-        ? { executionerTargetId: ts.roleState.executioner.targetId }
-        : {}),
-    };
-  }
-
-  if (myRole.id === WerewolfRole.TheThing) {
-    return extractTheThingState(game, callerId, nightActions, ts);
-  }
-
-  if (myRole.id === WerewolfRole.Insomniac) {
-    return extractInsomniacState(game, callerId, nightActions);
-  }
-
-  if (myRole.id === WerewolfRole.Count) {
-    return extractCountState(game, ts);
-  }
-
-  if (myRole.id === WerewolfRole.Arsonist) {
-    const arsonistAction = nightActions[myRole.id];
-    const soloAction =
-      arsonistAction && !isTeamNightAction(arsonistAction)
-        ? arsonistAction
-        : undefined;
-    return {
-      myNightTarget: soloAction?.skipped ? null : soloAction?.targetPlayerId,
-      myNightTargetConfirmed: soloAction?.confirmed ?? false,
-      ...(ts?.roleState?.arsonist?.dousedPlayerIds.length
-        ? { arsonistDousedPlayerIds: ts.roleState.arsonist.dousedPlayerIds }
-        : {}),
-    };
-  }
-
-  if (myRole.id === WerewolfRole.OneEyedSeer) {
-    const lockedTargetId = ts?.roleState?.oneEyedSeer?.lockedTargetId;
-    if (lockedTargetId && !ts.deadPlayerIds.includes(lockedTargetId)) {
-      return {
-        myNightTarget: undefined,
-        myNightTargetConfirmed: false,
-        oneEyedSeerLockedTargetId: lockedTargetId,
-      };
-    }
-  }
-
-  if (myRole.id === WerewolfRole.ElusiveSeer) {
-    if (ts?.turn === 1) {
-      const elusiveSeerVillagerIds = game.roleAssignments
-        .filter((a) => a.roleDefinitionId === (WerewolfRole.Villager as string))
-        .map((a) => a.playerId);
-      return {
-        myNightTarget: undefined,
-        myNightTargetConfirmed: false,
-        elusiveSeerVillagerIds,
-      };
-    }
-  }
-
-  if (myRole.id === WerewolfRole.Illuminati) {
-    const illuminatiAction = nightActions[myRole.id];
-    const soloAction =
-      illuminatiAction && !isTeamNightAction(illuminatiAction)
-        ? illuminatiAction
-        : undefined;
-    if (soloAction?.resultRevealed) {
-      const illuminatiRoleAssignments = game.roleAssignments.map((a) => {
-        const roleDef = getWerewolfRole(a.roleDefinitionId);
-        return {
-          playerId: a.playerId,
-          roleName: roleDef?.name ?? a.roleDefinitionId,
-          team: roleDef?.team ?? Team.Good,
-        };
-      });
-      return {
-        myNightTarget: undefined,
-        myNightTargetConfirmed: false,
-        illuminatiRoleAssignments,
-      };
-    }
-    return {
-      myNightTarget: undefined,
-      myNightTargetConfirmed: false,
-    };
-  }
-
-  if (myRole.id === WerewolfRole.EvilEmpath) {
-    const empathAction = nightActions[myRole.id];
-    const soloAction =
-      empathAction && !isTeamNightAction(empathAction)
-        ? empathAction
-        : undefined;
-    return {
-      myNightTarget: undefined,
-      myNightTargetConfirmed: soloAction?.confirmed ?? false,
-      ...(soloAction?.confirmed &&
-      ts?.roleState?.evilEmpath?.lastResult !== undefined
-        ? { evilEmpathNightResult: ts.roleState.evilEmpath.lastResult }
-        : {}),
-    };
-  }
-
-  return undefined;
-}
-function extractWitchState(
-  game: Game,
-  nightActions: Record<string, AnyNightAction>,
-  myRole: WerewolfRoleDefinition,
-  deadPlayerIds: string[],
-  ts: WerewolfTurnState | undefined,
-): Partial<WerewolfPlayerGameState> {
-  const witchAction = nightActions[myRole.id];
-  const soloAction =
-    witchAction && !isTeamNightAction(witchAction) ? witchAction : undefined;
-  const result: Partial<WerewolfPlayerGameState> = {
-    myNightTarget: soloAction?.skipped ? null : soloAction?.targetPlayerId,
-    myNightTargetConfirmed: soloAction?.confirmed ?? false,
-    witchAbilityUsed: ts?.roleState?.witch?.abilityUsed ?? false,
-  };
-  if (!ts?.roleState?.witch?.abilityUsed) {
-    const attacked = getInterimAttackedPlayerIds(
-      nightActions,
-      game.roleAssignments,
-      deadPlayerIds,
-      ts?.roleState?.priest?.wards,
-      ts?.roleState?.mirrorcaster?.charged,
-      ts?.roleState?.mercenary?.charged,
-      ts?.roleState?.arsonist?.dousedPlayerIds,
-    );
-    if (attacked.length > 0) {
-      result.nightStatus = attacked.map(
-        (id): NighttimeNightStatusEntry => ({
-          targetPlayerId: id,
-          effect: "attacked",
-        }),
-      );
-    }
-  }
-  return result;
-}
-
-function extractAltruistState(
-  game: Game,
-  callerId: string,
-  nightActions: Record<string, AnyNightAction>,
-  myRole: WerewolfRoleDefinition,
-  deadPlayerIds: string[],
-  ts: WerewolfTurnState | undefined,
-): Partial<WerewolfPlayerGameState> {
-  const altruistAction = nightActions[myRole.id];
-  const soloAction =
-    altruistAction && !isTeamNightAction(altruistAction)
-      ? altruistAction
-      : undefined;
-  const witchAction = nightActions[WerewolfRole.Witch] as
-    | { targetPlayerId?: string }
-    | undefined;
-  const witchProtectedId = witchAction?.targetPlayerId;
-  const attacked = getInterimAttackedPlayerIds(
-    nightActions,
-    game.roleAssignments,
-    deadPlayerIds,
-    ts?.roleState?.priest?.wards,
-    ts?.roleState?.mirrorcaster?.charged,
-    ts?.roleState?.mercenary?.charged,
-    ts?.roleState?.arsonist?.dousedPlayerIds,
-  ).filter((id) => id !== callerId && id !== witchProtectedId);
-  const result: Partial<WerewolfPlayerGameState> = {
-    myNightTarget: soloAction?.skipped ? null : soloAction?.targetPlayerId,
-    myNightTargetConfirmed: soloAction?.confirmed ?? false,
-  };
-  if (attacked.length > 0) {
-    result.nightStatus = attacked.map(
-      (id): NighttimeNightStatusEntry => ({
-        targetPlayerId: id,
-        effect: "attacked",
-      }),
-    );
-  }
-  return result;
-}
-
-function extractTheThingState(
-  game: Game,
-  callerId: string,
-  nightActions: Record<string, AnyNightAction>,
-  ts: WerewolfTurnState | undefined,
-): Partial<WerewolfPlayerGameState> {
-  const thingAction = nightActions[WerewolfRole.TheThing];
-  const soloAction =
-    thingAction && !isTeamNightAction(thingAction) ? thingAction : undefined;
-  const playerOrder = game.playerOrder ?? game.players.map((p) => p.id);
-  const adjacentPlayerIds = getAdjacentPlayerIds(playerOrder, callerId);
-  return {
-    myNightTarget: soloAction?.skipped ? null : soloAction?.targetPlayerId,
-    myNightTargetConfirmed: soloAction?.confirmed ?? false,
-    ...(adjacentPlayerIds.length > 0 ? { adjacentPlayerIds } : {}),
-    ...(ts?.roleState?.theThing?.tapped === callerId
-      ? { thingTappedMe: true }
-      : {}),
-    ...(soloAction?.confirmed && soloAction.targetPlayerId
-      ? { thingTappedPlayerId: soloAction.targetPlayerId }
-      : {}),
-  };
-}
-
-function extractInsomniacState(
-  game: Game,
-  callerId: string,
-  nightActions: Record<string, AnyNightAction>,
-): Partial<WerewolfPlayerGameState> {
-  const playerOrder = game.playerOrder ?? game.players.map((p) => p.id);
-  const { left, right } = getNeighborIds(playerOrder, callerId);
-
-  // Derive "woke and acted" purely from nightActions — nightPhaseOrder is empty
-  // during Daytime, so we cannot use it here. Any non-skip entry in nightActions
-  // for the neighbor's phase key means they woke and acted this night.
-  const neighborActed = (neighborId: string | undefined): boolean => {
-    if (!neighborId) return false;
-    const assignment = game.roleAssignments.find(
-      (a) => a.playerId === neighborId,
-    );
-    if (!assignment) return false;
-    const roleDef = getWerewolfRole(assignment.roleDefinitionId);
-    if (!roleDef) return false;
-    // Roles that wake with another role use that role's phase key.
-    const phaseKey = (roleDef.wakesWith ?? roleDef.id) as string;
-    const action = nightActions[phaseKey];
-    if (!action) return false;
-    if (isTeamNightAction(action)) {
-      // Group phase: check if the neighbor cast a non-skip vote.
-      const vote = action.votes.find((v) => v.playerId === neighborId);
-      return vote !== undefined && !vote.skipped;
-    }
-    // alerted: true (Veteran alert action) counts as acting even though it
-    // has no targetPlayerId.
-    return (
-      !action.skipped &&
-      (action.targetPlayerId !== undefined || action.alerted === true)
-    );
-  };
-
-  return {
-    myNightTarget: undefined,
-    myNightTargetConfirmed: false,
-    insomniacResult: {
-      leftActed: neighborActed(left),
-      rightActed: neighborActed(right),
-    },
-  };
-}
-
-function extractCountState(
-  game: Game,
-  ts: WerewolfTurnState | undefined,
-): Partial<WerewolfPlayerGameState> {
-  // The Count only acts on night 1.
-  if (ts?.turn !== 1) {
-    return {
-      myNightTarget: undefined,
-      myNightTargetConfirmed: false,
-    };
-  }
-
-  const playerOrder = game.playerOrder ?? game.players.map((p) => p.id);
-  // Exclude the narrator (ownerPlayerId) from the seating list.
-  const seatOrder = playerOrder.filter((id) => id !== game.ownerPlayerId);
-  const mid = Math.ceil(seatOrder.length / 2);
-  const leftHalf = seatOrder.slice(0, mid);
-  const rightHalf = seatOrder.slice(mid);
-
-  const isWerewolfPlayer = (playerId: string): boolean => {
-    const assignment = game.roleAssignments.find(
-      (a) => a.playerId === playerId,
-    );
-    const roleDef = assignment
-      ? getWerewolfRole(assignment.roleDefinitionId)
-      : undefined;
-    return roleDef?.isWerewolf === true;
-  };
-
-  return {
-    myNightTarget: undefined,
-    myNightTargetConfirmed: false,
-    countResult: {
-      leftCount: leftHalf.filter(isWerewolfPlayer).length,
-      rightCount: rightHalf.filter(isWerewolfPlayer).length,
-    },
   };
 }
